New HTML sandbox formfield based on the Monaco editor formfield
This field implement an HTML sandbox. The user can edit HTML code and see the result rendered in real time as a standalone document using an iframe. This formfield is meant for quick prototyping of HTML pages which can then be shared among other wiki users, taking also advantage of other wiki features such as versioning, access control and so on.